LIBCIR-323. Added DSpace ORCID Authority functionality
Updated the configuration as described in

https://wiki.lyrasis.org/display/DSDOC7x/ORCID+Authority

to enable the DSpace ORCID Authority functionality.

The "local.cfg.EXAMPLE" configuration file is configured to use the
ORCID sandbox API endpoint, as it does not appear to require ORCID
credentials.

Added "dspace/docs/OrcidIntegration.md" file with information about
the ORCID integration in MD-SOAR.

https://umd-dit.atlassian.net/browse/LIBCIR-323